backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==25.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
